Add SparsityConfig class for PyTorch inference
