+++ intel_extension_for_transformers/transformers/utils/utility.py
@@ -30,6 +30,8 @@ ENCODER_NAME = "encoder_model.bin"
 DECODER_NAME = "decoder_model.bin"
 DECODER_WITH_PAST_NAME = "decoder_with_past_model.bin"
 WEIGHTS_NAME = "pytorch_model.bin"
+QUANT_CONFIG = "quantization_config.json"
+SPARSITY_CONFIG = "sparse_config.json"
 
 torch = LazyImport("torch")
